+++ lib/money/ecto/money_ecto_composite_type.ex
@@ -8,20 +8,28 @@ if Code.ensure_loaded?(Ecto.Type) do
     amount is stored as a decimal number,
     """
 
-    @behaviour Ecto.Type
+    use Ecto.ParameterizedType
 
-    def type do
+    def type(_params) do
       :money_with_currency
     end
 
-    def blank?(_) do
-      false
+    def init(opts) do
+      opts
+      |> Keyword.delete(:field)
+      |> Keyword.delete(:schema)
     end
 
     # When loading from the database
-    def load({currency, amount}) do
+    def load(tuple, loader \\ nil, params \\ [])
+
+    def load(nil, _loader, _params) do
+      {:ok, nil}
+    end
+
+    def load({currency, amount}, _loader, params) do
       with {:ok, currency_code} <- Money.validate_currency(currency) do
-        {:ok, Money.new(currency_code, amount)}
+        {:ok, Money.new(currency_code, amount, params)}
       else
         _ -> :error
       end
@@ -30,70 +38,71 @@ if Code.ensure_loaded?(Ecto.Type) do
     # Dumping to the database.  We make the assumption that
     # since we are dumping from %Money{} structs that the
     # data is ok
-    def dump(%Money{} = money) do
+    def dump(money, dumper \\ nil, params \\ [])
+
+    def dump(%Money{} = money, _dumper, _params) do
       {:ok, {to_string(money.currency), money.amount}}
     end
 
-    def dump(_) do
+    def dump(_, _, _) do
       :error
     end
 
     # Casting in changesets
-    def cast(%Money{} = money) do
+    def cast(money, params \\ [])
+
+    def cast(%Money{} = money, _params) do
       {:ok, money}
     end
 
-    def cast(%{"currency" => _, "amount" => ""}) do
+    def cast(%{"currency" => _, "amount" => ""}, _params) do
       {:ok, nil}
     end
 
-    def cast(%{"currency" => currency, "amount" => amount})
+    def cast(%{"currency" => currency, "amount" => amount}, params)
         when (is_binary(currency) or is_atom(currency)) and is_integer(amount) do
       with decimal_amount <- Decimal.new(amount),
            {:ok, currency_code} <- Money.validate_currency(currency) do
-        {:ok, Money.new(currency_code, decimal_amount)}
+        {:ok, Money.new(currency_code, decimal_amount, params)}
       else
         {:error, {_, message}} -> {:error, message: message}
       end
     end
 
-    def cast(%{"currency" => currency, "amount" => amount})
+    def cast(%{"currency" => currency, "amount" => amount}, params)
         when (is_binary(currency) or is_atom(currency)) and is_binary(amount) do
       with {amount, ""} <- Cldr.Decimal.parse(amount),
            {:ok, currency_code} <- Money.validate_currency(currency) do
-        {:ok, Money.new(currency_code, amount)}
+        {:ok, Money.new(currency_code, amount, params)}
       else
         {:error, {_, message}} -> {:error, message: message}
         :error -> {:error, message: "Couldn't parse amount #{inspect amount}"}
       end
     end
 
-    def cast(%{"currency" => currency, "amount" => %Decimal{} = amount})
+    def cast(%{"currency" => currency, "amount" => %Decimal{} = amount}, params)
         when is_binary(currency) or is_atom(currency) do
       with {:ok, currency_code} <- Money.validate_currency(currency) do
-        {:ok, Money.new(currency_code, amount)}
+        {:ok, Money.new(currency_code, amount, params)}
       else
         {:error, {_, message}} -> {:error, message: message}
       end
     end
 
-    def cast(%{currency: currency, amount: amount}) do
-      cast(%{"currency" => currency, "amount" => amount})
+    def cast(%{currency: currency, amount: amount}, params) do
+      cast(%{"currency" => currency, "amount" => amount}, params)
     end
 
-    def cast(string) when is_binary(string) do
-      case Money.parse(string) do
+    def cast(string, params) when is_binary(string) do
+      case Money.parse(string, params) do
         {:error,{_, message}} -> {:error, message: message}
         money -> {:ok, money}
       end
     end
 
-    def cast(_money) do
+    def cast(_money, _params) do
       :error
     end
 
-    # New for ecto_sql 3.2
-    def embed_as(_), do: :self
-    def equal?(term1, term2), do: term1 == term2
   end
 end