+++ priv/repo/migrations/20500930144804_create_test_table.exs
@@ -6,6 +6,7 @@ defmodule Money.Repo.Migrations.CreateMoneyTable do
       add :name,            :string
       add :employee_count,  :integer
       add :payroll,         :money_with_currency
+      add :tax,             :money_with_currency
       timestamps()
     end
   end