+++ test/support/schema.ex
@@ -4,6 +4,7 @@ defmodule Organization do
   @primary_key false
   schema "organizations" do
     field :payroll, Money.Ecto.Composite.Type
+    field :tax, Money.Ecto.Composite.Type, fractional_digits: 4
     field :name,            :string
     field :employee_count,  :integer
     timestamps()