+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# Changelog for Money_SQL v1.4.0
+
+This is the changelog for Money_SQL v1.4.0 released on ____, 2021.
+
+### Enhancements
+
+* Changes `Money.Ecto.Composite.Type` and `Money.Ecto.Map.Type` to be `ParameterizedType`. As a result, Ecto 3.5 or later is required. This change allows configuration of format options for the `:money_with_currency` to added as parameters in the Ecto schema.  For the example schema:
+```elixir
+defmodule Organization do
+  use Ecto.Schema
+
+  @primary_key false
+  schema "organizations" do
+    field :payroll,         Money.Ecto.Composite.Type
+    field :tax,             Money.Ecto.Composite.Type, fractional_digits: 4
+    field :name,            :string
+    field :employee_count,  :integer
+    timestamps()
+  end
+end
+```
+The field `:tax` will be instantiated as a `Money.t` with `:format_options` of `fractional_digits: 4`.
+
 # Changelog for Money_SQL v1.3.1
 
 This is the changelog for Money_SQL v1.3.1 released on September 30th, 2020.